Update seeds file
Before this commit, I was unable to start the project. This was because we were shoveling User.first into the admin.users method, but a user did not exist.

Now we are creating a user. You will need to put your IU assigned email and password in the .env file.