Remove files that are covered by IIIF Print
This commit will remove iiif_helper, _index_header_list_default, and
_universal_viewer. These files are covered by the IIIF Print gem and
would conflit if they were left in the application.  This added a
configuration to the initializer to set the UV config path to account
for `uv_config.json` (with an underscore and not a -).  Lastly, we had
to change the config/uv/uv.html to account for the removal of the
javascript and jquery that was in the _universal_viewer partial to
maintain the same functionality of doing an automatic search from the
catalog to UV.