Remove IIIF Print implementation of OCR
This commit will remove the IIIF Print implementation of OCR generation.
It is being reverted to Essi's original implementation but wiring up the
UV search with IIIF Print's version.

ESSI::BlacklightIiifSearch::AnnotationBehavior is not being used and
instead we're using IiifPrint::BlacklightIiifSearch::AnnotationDecorator
for UV search.

The key difference in implementation is that while Essi searches an
indexed field, IIIF Print searches a json file.