Remove unnecessary spec and add configuration
This commit will remove the pluggable derivative spec that is no longer
relevant. It will also add the configuration for adding the all_text_*
index fields from the alto xml that Essi already implements.

This commit will also leverage another configuration option to allow the
application to specify where to pull the json coordinates from.  Essi
holds the json coordinates in the `word_boundary_tsi` field of the
SolrDocument.